优化 collect.py: 异步IO、按场景分组、去重、depth使用float16节省空间
主要改动:
- 添加去重逻辑：扫描已采集的episode_id，跳过重复
- 按场景分组采集：减少场景切换次数，提升速度
- 异步IO：使用线程池并行保存RGB和Depth
- RGB使用JPEG格式（质量95%）
- Depth使用float16格式，空间减半
- 只在场景变化时重建ShortestPathFollower

+++ collect.py
@@ -713,7 +713,7 @@ while clip_id <= NUM_CLIPS:
             depth_images.append(depth.copy())
             # 🚀 异步保存 depth
             depth_path = depth_dir / f"{frame_id:06d}.npy"
-            io_futures.append(executor.submit(np.save, str(depth_path), depth.copy()))
+            io_futures.append(executor.submit(np.save, str(depth_path), depth.copy().astype(np.float16)))  # float16 节省空间
 
         agent_state = sim.get_agent_state()
         trajectory_positions.append(agent_state.position.copy())
@@ -763,7 +763,7 @@ while clip_id <= NUM_CLIPS:
                 # 🚀 异步保存 depth
                 depth_path = depth_dir / f"{frame_id:06d}.npy"
                 depth_copy = depth.copy()
-                io_futures.append(executor.submit(np.save, str(depth_path), depth_copy))
+                io_futures.append(executor.submit(np.save, str(depth_path), depth_copy.astype(np.float16)))  # float16 节省空间
 
             if clip_failed:
                 break
